+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [0.1.4] (2025-12-15)
 ### Changed
 * cleanup backends
 
@@ -13,7 +16,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * a `user_id` column into a `Logs` table
 * a `ipaddr` column into a `Logs` table
 * tests
-
 
 ## [0.1.3] (2025-12-11)
 ### Added
@@ -27,7 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * first commit
 
-[Unreleased]: https://github.com/aki-akaguma/browserinfocm/compare/v0.1.3..HEAD
+[Unreleased]: https://github.com/aki-akaguma/browserinfocm/compare/v0.1.4..HEAD
+[0.1.4]: https://github.com/aki-akaguma/browserinfocm/compare/v0.1.3..v0.1.4
 [0.1.3]: https://github.com/aki-akaguma/browserinfocm/compare/v0.1.2..v0.1.3
 [0.1.2]: https://github.com/aki-akaguma/browserinfocm/compare/v0.1.1..v0.1.2
 [0.1.1]: https://github.com/aki-akaguma/browserinfocm/compare/v0.1.0..v0.1.1